Find special divisor
some special numbers.one of such property we will try to  find in this problem. Some number have a special divisor or factor F (except the number itself) that ultimately

generate quotient 1 and remainder 0 if we divide number N by factor F ,m times. The process is as follows

N / F =Q1 and R1=0 (where N is Dividend, F is Divisor,  Q1 is Quotient and R1 is Remainder  )

Q1 / F= Q2 and R2=0(Now Q1 is Dividend but the Divisor is same F , Q2 is new Quotient and R2 is Remainder )