Rearrangement array
Your task is to create a program that accepts an array A of N unique positive integers. Now find the first N/2 greatest elements G(i) in array A and arrange them in first N/2 cells of Array A in decreasing order. after this find the rest N/2 smallest elements S(i) of array A and place them in rest N/2 cells of array A in ascending order.

test case:
10 20 5 1 30 12

30 20 10 1 5 12