Add a printNearestFibonicciNumber(int);
For example, if N is 10, N is closest  to F(6)=8 rather than F(7)=13 because there are less number between N and F(6) rather than N and F(7). (exclude N and F(i))

If given number is equally closest to 2 Fibonacci numbers of Fibonacci sequence print both numbers in increasing order in output.

For example if N is  4 then it is closest to Fibonacci numberF(5)=3 and F(6)=5 because there are zero numbers available between N and F(5) and also between N and F(6) (exclude N and F(i)). so print F(5) and F(6) both in output and should be in increasing order that is 3 5.

Example:
6
5