Added int reduceToZeroByGivenNumber(int *arr, int n, int x); function
Statement: create an array of size n and accept n integers in it. now accept one more integer x denoting a random value.

Your Task is to reduce the x to zero by subtracting array elements from this number. remember you should subtract minimum elements to

make x zero.

for example: suppose array elements are 1 3 2 1 4 2 5 4 and  x=6 so if you subtract 6 by 5 and 1 or 4 and 2 in both cases you subtracted 2

elements from x and if you subtract x by 3 2 and 1 you subtracted 3 elements to reduce x to zero.

the constraint is you should subtract minimum elements to make x zero.

so output contains minimum elements as 2.

if it is not possible to reduce x to zero print -1 in output