Added countSwapOperation(char *s, char *p);
Statement:if two string P and Q of same length N, contains same set of characters C(i) but in different order they are called as anagram of each other.

For example, listen and silent are anagram of each other. Ashok was thinking to converts Q into P by using sequence of some swap operations.

for this he creates an algorithm which is as follows

Step 1: look at the position i=1 of string P and Q. it might happened that P and Q both having same character C on position i=1. But there is also a

possibility that on i=1position both P and Q have different characters.

Step 2: If P and Q have same character on position i=n (1<=n<=N) repeat the process for i=i+1position.

Step 3: If P have character C on position i=1 and Q have character D on position i=1 then find character C in String Q which might find on position i=n in

String Q so swap characters C and D in String Q.

Step 4: Now repeat the above steps for i=2 to n position and rearrange the Q into P.

Your task is to write a program to accept two string P and Q of size N which is anagram of each other and convert Q in to P using above algorithm at the time

of conversion your program should count the swap operation S that are performed over Q to make it P. finally print the number of swap operation S.