Added void printPpythagoreanTriplets(int final);
Print Pythagorean triplets available in a range 

Statement:Any three numbers X, Y and Z are called Pythagorean triplet if they satisfy following conditions

X<Y<Z

and X^2+Y^2=Z^2. (X^2 read as X square)

In order to print Pythagorean triplets in a given range accept an integer N and treat this N as an upper limit of the range R which starts from1 where N>=Z.

Your task is to write a program to print all Pythagorean triplets available in that range

Note:suppose N=25 then we can find 3 Pythagorean triplets in between range of 1 to 25 (where 1 and 25 are inclusive). In order to make this as a valid test case you

should print them in increasing order of there availability that means print triplets in increasing order of X (first number of triplet) associated with each triplet.

Example
input -> 25
output 
3 4 5
5 12 13
6 8 10
7 24 25
8 15 17
9 12 15
12 16 20
15 20 25