Added int isArrayHasSubset(int *, int *, int, int)
Is an array is subset of another array? 

Statement:Array subset is defined as small part of array which consist consecutive array elements. Any subset S of array A can be formed as one or more than one

adjacent cells of array  A. so a single element A[i] (where i represents index of array) or group of consecutive elements of array A[i] A[i+1] A[i+2] …… can be called as

subset of array A.In this program you have to compare two arrays A and B. in order to find that B is subset of array A or not.

For this create a program such that it accepts two arrays A and B that contains M and N elements respectively and find that array B is subset of array A or not. if B is

subset of A print 1 in output and if B is not subset of A print -1 in output.

Input:
10
1 2 3 4 5 6 7 8 9 10
3
8 9 10
                     
Expected:   1

Input:
9
5 6 9 21 54 97 23 67 2
4
21 54 23 67
Expected:   -1