Updated admission();
Statement:There are 3 popular courses  CR offered by the college named as X, Y and Z. each course have limited seats so college management can’t give the admissions to all

applied student for certain course. Course X, Y and Z requires some special skills in certain subjects so college management decided to make an admission process based on

the skills required for all 3 courses.in order to give the admission in any course they accept marks in Three Subjects A, B and C. now on the basis of these subjects marks they

check the student is eligible for which course or doesn’t eligible for any course. Following is the criteria to check the eligibility of student.

In order to take admission in X course the required marks in subjects A,B and C and total marks in A+B+C are as follows

A>=90

B>=85

C>=75

A+B+C>=255

In order to take admission in Y course the required marks in subjects A,B and C  and total marks in A+B+C  are as follows

A>=80

B>=90

C>=85

A+B+C>=260

In order to take admission in Z course the required marks in subjects A,B and Cand total marks in A+B+Care as follows

A>=75

B>=80

C>=85

A+B+C>=255

If student doesn’t fit in any criteria mentioned above, he is not eligible for admission.

The admission process is flexible in terms of choosing the course so student will enter his favorite course name in which he wants to take admission along with marks in

subject A, B and C.

Your task is to write a program that accepts marks AM, BM and CM denoting marks of subjects A, B and C respectively. Also accepts the Course X, Y or Z of student’s choice.

Find the total of AM+BM+CM and tell him he is eligible for that course or not.

Also, if student is not eligible for his favorite course in which he wants to take admission your program should give him a suggestion that, is he eligible for another course.

If he is eligible for his favorite course print 1 along with course name separated by space in a line.

If he is not eligible for his favorite course but eligible for different course print -1 along with the course name and in next line print 1 along with the Course name in which he is eligible.

If he is not at all eligible for any course print 0.

Input:first line contains three integers  AM BM CM denoting marks in three subjects A, B and C respectively andan upper-case character CR denoting favorite course of student. (possible values are X, Y and Z) separated by space.

Output:first line either contain 1 or -1 along with favorite course name CR separated by space or first line contains 0 if student is not eligible.

If first line contains -1 second line will contain 1 along with Course name CR in which he/she is eligible.

Constraints:0<=AM<=100

0<=BM<=100

0<=CM<=100

CR can be either X, Y or Z only.

Example:

80 85 90 Y

-1 Y

1 Z